pyproject.toml
[tool.poetry]
name = "pythonproject"
version = "0.1.0"
description = ""
authors = ["rednek21 <egor.podustov@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.7"
djangorestframework = "^3.15.2"
channels = "^4.1.0"
channels-redis = "^4.2.0"
daphne = "^4.1.2"
django-redis = "^5.0.0"
python-environ = "^0.4.54"
psycopg2-binary = "^2.9.9"

[tool.poetry.dev-dependencies]
flake8 = "^6.0"
isort = "^5.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
